pushed

+++ env.sh
@@ -1,30 +1,27 @@
 #!/bin/bash
 
-[[ "${BASH_SOURCE[0]}" == "${0}" ]] && {
-    echo "Please use 'source' or '.' to run env.sh."
-}
-
-if [[ "$(uname)" != "Darwin" ]]; then
-    python3 -m venv --help >/dev/null 2>&1 || {
-        echo "Installing python3-venv..."
-        sudo apt install -y python3-venv
-    }
+if [[ "${BASH_SOURCE[0]}" == "${0}" ]]; then
+    echo "Please use 'source' or '.' to run $(basename "$0")."
+    return 1
 fi
 
-[ ! -d env ] && {
+if [[ "$(uname)" != "Darwin" ]] && ! python3 -m venv --help &>/dev/null; then
+    echo "Installing python3-venv..."
+    sudo apt update && sudo apt install -y python3-venv
+fi
+
+if [[ ! -d env ]]; then
     echo "Creating virtual environment..."
     python3 -m venv env
-    
+fi
+
+if [[ -z "$VIRTUAL_ENV" ]]; then
     echo "Activating virtual environment..."
     source env/bin/activate
+fi
 
-    [ -f requirements.txt ] && {
-        echo "Installing requirements..."
-        pip install -r requirements.txt
-    }
-}
-
-[ -z "$VIRTUAL_ENV" ] && {
-    echo "Activating virtual environment..."
-    source env/bin/activate
-}
+if [[ -f requirements.txt ]]; then
+    echo "Installing requirements..."
+    pip install --upgrade pip
+    pip install -r requirements.txt
+fi
